requirements.txt
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
djangorestframework==3.10.0
djangorestframework-simplejwt==3.2.3
gunicorn==20.0.4
psycopg2==2.8.6
PyJWT==1.6.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2018.5
PyYAML==5.3.1
sqlparse==0.4.1
whitenoise==5.2.0
django-cors-headers
dj-static==0.0.6
django-toolbelt
opencv-python
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0+cpu
